requirements.txt
asgiref==3.2.3
crcmod==1.7
Django==3.0.3
pycrypto==2.6.1
pytz==2019.3
sqlparse==0.3.0
